Commit on 18072017

+++ MNTR/src/main/java/com/cbo/mntr/constants/UserConstants.java
@@ -12,7 +12,7 @@ public class UserConstants implements Serializable {
 	public static final int ldapLogin = 2;
 
 	public static final String defaultLogoPathKey = "user.logo.path";
-	public static final int anonymousUser = 0;
+	public static final Long anonymousUser = 0L;
 	public static final String systemUserRole = "SYSTEM USER ROLE";
 	public static final String suaCount = "suaCount";
